Move assert_chunks_equal to xarray.testing.
